Fix: API returns Double for rollingFiveHourLimit.remaining, not Int (#43)

+++ android/app/src/main/java/com/duchastel/simon/syntheticwidget/data/QuotaModels.kt
@@ -18,7 +18,7 @@ data class RollingFiveHourLimit(
     @SerialName("tickPercent")
     val tickPercent: Double,
     @SerialName("remaining")
-    val remaining: Int,
+    val remaining: Double,
     @SerialName("max")
     val max: Int,
     @SerialName("limited")
